requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
reportlab>=3.6.0
plotly>=5.0.0
bcrypt>=4.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0